Added calls to other endpoints of the api on the web store api client.
Added calls to GET quote and POST quote items.

Added logic to handle cookies for the quote endpoints. The cookies are how the API identifies the quote for subsequent requests.

Added logic to instantiate and configure a new http client obejct that is aware of cookies. Use that for the quote endpoint calls intead of the default dependency injected one.

Added more fields to the product class to represent response payloads from the API.