Create checkout payment information page.
Added structure and styling for the page.

Implemented integration with Authorize.NET.

Added backend logic to redirect to the cart page if trying to access this page with no quote.

Added backend logic to validate the quote by calling the API endpoint for it.

Added backend logic to post the order to the API.

Added Api client methos to call POST quote validate and POST order endpoints.

Link to the checkout payment page from the checkout shipping page.