Render order details on the success page.
Developed markup to render the order. Added more fields to the Order API DTO so that they template can render it.

Store the order as JSON in the session after the order is placed in the payment page, so that the success page can retrieve it and render it.

Clear the API quote cookie from the session after the order is placed.

Use property with get and set instead of separate Get* and Set* methods for storing and retrieving the quote cookie from session. Update code that uses it accordingly.

Move the states list of options and fetch method to the base page model.