Fix various issues with the account email confirmation feature.
Include the user id as a query string param in the confirm email link. This needs to be passed to the ASP.NET Identity API endpoint.

Fixed generation of email confirmation code by applying Base 64 encode. The raw return value of the GenerateEmailConfirmationTokenAsync method cannot be used as is. So that was producing links that did not work.

HTML decode the link in order confirmation emails so that it is properly rendered in the email template when run by the default ASP.NET Identity API register endpoint. For some reason the framework sends in a URL escaped & in the query string instead of the literal value. This is not as important because this endpoint is not supposed to be used in this API. We have the POST Users endpoint for that. But still, its better for it to not be broken and the fix is cheap enough.