Implemented the forgot password flow.
Added methods to the API client class to call the forgot password and reset password endpoints.

Added forgot password link to the loginc page.

Added forgot passsword page which allows requesting a password reset email by calling the API endpoint.

Add a reset password page, which the email links to, that can be used to change a password by calling the API endpoint.