Fix issue with using MapStaticAssets instead of UseStaticFiles
Had to adjust the dockerfiles for this to work. That is, added pnpm install and pnpm run build commands right before the dotnet publish command.

The reason why MapStaticAssets was not working as things were is because that function depends on the static files being there well before build time. Running these commands in the dockerfile right before dotnet publish makes it so. It was not enough running these commands as part of the MSBuild process, as defined in the csproj files.

Anyway, it is still convenient to have the commands in the csproj file for development purposes. That way it is not necessary to manually run them every time a change in js or css is made. So I left it there.

Still, made an improvement there by changing the BeforeTargets to "BeforeBuild" instead of just "Build". This makes sure to run the commands earlier.

(cherry picked from commit 3bcf0e67eb4232cf480f149228359b0616555fab)